Slight refactor of pages/index
