+++ components/UnifiedDocFeed/utils/getFetchDefaults.ts
@@ -1,0 +1,14 @@
+export const getFetchDefaults = ({ query, authToken }) => {
+  const defaultProps = {
+    feed: 0,
+    home: true,
+    initialFeed: null,
+    initialHubList: null,
+    leaderboardFeed: null,
+    loggedIn: authToken !== undefined,
+    page: 1,
+    query,
+  };
+
+  return defaultProps;
+}
